add the function to select、update and delete radar_data

+++ big_angle_radar/big_angle_radar/sql/create_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS radar_data (
     id integer PRIMARY KEY AUTOINCREMENT, 
-    targetid integer,
+    targetid float,
     x float,
     y float,
     speed float,
